% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{entries}
\alias{entries}
\alias{entries,MetaboliteDatabase-method}
\title{Retrieve entries}
\usage{
entries(db)

\S4method{entries}{MetaboliteDatabase}(db)
}
\arguments{
\item{db}{MetaboliteDatabase}
}
\description{
return accession data table from MetaboliteDatabase object
}
