% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{adductTransformMF}
\alias{adductTransformMF}
\title{Molecular formula adduct transformation}
\usage{
adductTransformMF(MF, adduct, adduct_rules_table = adduct_rules())
}
\arguments{
\item{MF}{a molecular formula to transform}

\item{adduct}{a transformation adduct}

\item{adduct_rules_table}{adduct formation rules. Defaults to \code{adduct_rules()}.}
}
\value{
A transformed molecular formula.
}
\description{
Transform a molecular formula based on a specified adduct.
}
\examples{
adductTransformMF('C6H12O6','[M+H]1+')
}
