% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acids.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{Example amino acids}
\format{
a tibble containing 20 rows and 6 columns
\describe{
\item{ACCESSION_ID}{}
\item{NAME}{}
\item{InChi}{}
\item{InChiKey}{}
\item{SMILES}{}
}
}
\usage{
amino_acids
}
\description{
Chemical structures of the essential amino acids
}
\keyword{datasets}
