% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotation of FIE-HRMS m/z}
\usage{
annotate(explan.mass, peak.mat, dat.all, mode, ppm, MZedDB, Path, lat = c(C =
  90, iC = 1, H = 180, iH = 0, N = 10, iN = 0, O = 40, F = 0, Na = 1, Si = 0, P
  = 6, S = 5, Cl = 0, iCl = 0, Br = 0, iBr = 0, K = 1, iK = 1, iO = 1))
}
\description{
Annotation of FIE-HRMS m/z
}

