% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyIonisationRule.R
\name{applyIonisationRule}
\alias{applyIonisationRule}
\title{applyIonisationRule}
\usage{
applyIonisationRule(ID, adduct, adducts = mzAnnotation::Adducts,
  DB = mzAnnotation::MZedDB)
}
\arguments{
\item{ID}{database ID of the metabolite  forr which to apply formation rule}

\item{adduct}{adduct of which to apply rule}

\item{adducts}{adduct table containing available adduct rules}

\item{DB}{Database to use. Defaults to \code{MZedDB}}
}
\description{
apply an adduct formation rule to a database entry
}
\examples{
applyIonisationRule('D27880',"[M-H]1-")
}
