% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTransformations.R
\name{availableTransformations}
\alias{availableTransformations}
\title{availableTransfromations}
\usage{
availableTransformations()
}
\description{
returns transformations available in mzAnnotation
}
