% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcM}
\alias{calcM}
\title{Calculate the mass of an m/z}
\usage{
calcM(
  mz,
  adduct = "[M+H]1+",
  isotope = NA,
  transformation = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules(),
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{mz}{m/z for which to calculate M}

\item{adduct}{adduct to apply}

\item{isotope}{isotope to apply}

\item{transformation}{transformation to apply}

\item{adduct_rules_table}{tibble containing available adduct formation rules. Defaults to \code{adduct_rules()}.}

\item{isotope_rules_table}{tibble containing available isotopic rules. Defaults to \code{isotope_rules()}.}

\item{transformation_rules_table}{tibble containing available transformation rules. Defaults to \code{transformation_rules()}.}
}
\description{
calculate M for a given m/z, adduct, isotope and transformation
}
\examples{
calcM(118.08626,adduct = '[M+H]1+',isotope = '13C',transformation = 'M - [O] + NH2]')
}
