% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMZ.R
\name{calcMZ}
\alias{calcMZ}
\title{Calculate the m/z of a mass}
\usage{
calcMZ(
  M,
  adduct = "[M+H]1+",
  isotope = NA,
  transformation = NA,
  adductTable = adducts(),
  isotopeTable = isotopes(),
  transformationTable = transformations()
)
}
\arguments{
\item{M}{M for which to calculate an m/z}

\item{adduct}{adduct to apply}

\item{isotope}{isotope to apply}

\item{transformation}{transformation to apply}

\item{adductTable}{adduct table containing available adduct rules. Defaults to \code{adducts()}.}

\item{isotopeTable}{isotope table containing available isotope rules. Defaults to \code{isotopes()}.}

\item{transformationTable}{transformations table containing available transformations rules. Defaults to \code{transformations()}.}
}
\description{
calculate an m/z for a given M, adduct, isotope and transformation
}
\examples{
calcMZ(116.05182,adduct = '[M+H]1+',isotope = '13C',transformation = 'M - [O] + NH2]')
}
