% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{chemicalDescriptors}
\alias{chemicalDescriptors}
\title{Chemical descriptors}
\usage{
chemicalDescriptors(SMILES)
}
\arguments{
\item{SMILES}{a character vector of valid SMILES}
}
\description{
Chemical descriptors
}
\examples{
data(amino_acids)
chemicalDescriptors(amino_acids$SMILES)
}
