% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{CHOproportion}
\alias{CHOproportion}
\title{The proportion of C, H and O atoms in molecular formulas}
\usage{
CHOproportion(element_frequencies)
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble contianing the
}
\description{
Calculate the proportion of C, H and O in molecular formulas.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  CHOproportion()
}
