% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(input, inputType, outputType)
}
\arguments{
\item{input}{vecotr of valid SMILES or InChIs}

\item{inputType}{either "smiles" or "inchi", denoting the input type}

\item{outputType}{either "smiles", "inchi" or "inchikey", denoting the output type}
}
\description{
convert between SMILES and InChI and to InChIKey
}
\examples{
convert(aminoAcids$SMILES,'smiles','inchi')
}
