% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSmiles.R
\name{ConvertSmiles}
\alias{ConvertSmiles}
\title{Convert SMILES to Molecular Formula's and Exact Masses}
\usage{
ConvertSmiles(smile)
}
\arguments{
\item{smile}{a character string of a valid SMILE}
}
\value{
a \code{tibble} of Molecular Formula and Exact Mass
}
\description{
Convert SMILES to Molecular Formula's and Exact Masses
}
\examples{
ConvertSmile('C[C@H](C(=O)O)N')
}
