% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corLists.R
\name{corLists}
\alias{corLists}
\title{Aggregate significant correlations into lists for each variable}
\usage{
corLists(x)
}
\description{
Aggregate significant correlations into lists for each variable
}

