% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{elementCountCheck}
\alias{elementCountCheck}
\title{Element count checks}
\usage{
elementCountCheck(element_frequencies)
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble containing results of the element count checks. The column headers specify the check criteria. \code{TRUE} is returned where relevant checks
are passed and \code{FALSE} where not. \code{NA} is returned where the check is not relevant.
}
\description{
Element count checks based on Rule 6 of the Seven Golden Rules by Kind et al 2007.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  elementCountCheck()
}
\references{
Kind, T. and Fiehn, O., 2007. Seven Golden Rules for heuristic filtering of molecular formulas obtained by
accurate mass spectrometry. \emph{BMC bioinformatics}, \emph{8}(1), pp.1-20.
}
