% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{elementFrequencies}
\alias{elementFrequencies}
\title{Extract element frequencies}
\usage{
elementFrequencies(MF)
}
\arguments{
\item{MF}{a vector of molecular formulas}
}
\value{
A tibble containing element frequencies.
}
\description{
Extract the element frequencies from a molecular formula.
}
\examples{
elementFrequencies(c('H2O','C12H22O11'))
}
