% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{elementCountCheck}
\alias{elementCountCheck}
\title{Element probability check}
\usage{
elementCountCheck(element_frequencies)
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble containing results of the heteroatom ratio checks. The column headers refer to elemental counts needed to apply the heuristic rule. See Table 3 in Kind et al 2007 for definitions of the heuristic rules applied.
}
\description{
Element probability checks based on rule #6 of the Seven Golden Rules by Kind et al 2007.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  elementCountCheck()
}
