% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{filterACCESSIONS,MetaboliteDatabase-method}
\alias{filterACCESSIONS,MetaboliteDatabase-method}
\title{filterACCESSIONS}
\usage{
\S4method{filterACCESSIONS}{MetaboliteDatabase}(db, ids)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{ids}{vector of accession IDs}
}
\description{
Filter a MetaboliteDatabase based on given accession IDs.
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids))
db <- filterACCESSIONS(db,c(1,2))
}
