% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/filter-methods.R
\name{filterACCESSIONS}
\alias{filterACCESSIONS}
\alias{filterACCESSIONS,MetaboliteDatabase-method}
\title{Filter accessions}
\usage{
filterACCESSIONS(db, ids)

\S4method{filterACCESSIONS}{MetaboliteDatabase}(db, ids)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{ids}{vector of accession IDs}
}
\description{
Filter a MetaboliteDatabase based on given accession IDs.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
db <- filterACCESSIONS(db,c(1,2))
}
