% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{filterER,MetaboliteDatabase-method}
\alias{filterER,MetaboliteDatabase-method}
\title{filterER}
\usage{
\S4method{filterER}{MetaboliteDatabase}(db, rule)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{rule}{elemental rule given as a string}
}
\description{
Filter a MetaboliteDatabase based on an elemental rule.
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids))
db <- filterER(db,'S>0')
}
