% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/filter-methods.R
\name{filterER}
\alias{filterER}
\alias{filterER,MetaboliteDatabase-method}
\title{Filter using an elemental rule}
\usage{
filterER(db, rule)

\S4method{filterER}{MetaboliteDatabase}(db, rule)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{rule}{elemental rule given as a string}
}
\description{
Filter a MetaboliteDatabase based on an elemental rule.
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids$SMILES))
db <- filterER(db,'S>0')
}
