% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{filterIP,MetaboliteDatabase-method}
\alias{filterIP,MetaboliteDatabase-method}
\title{filterIP}
\usage{
\S4method{filterIP}{MetaboliteDatabase}(db, rule)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{rule}{Character containing ionisation rule.}
}
\description{
Filter MetaboliteDatabase based on an ionisation rule
}
\examples{
rule <- Adducts$Rule[52]
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids))
db <- filterIP(db,rule)
}
