% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{filterMR,MetaboliteDatabase-method}
\alias{filterMR,MetaboliteDatabase-method}
\title{filterMR}
\usage{
\S4method{filterMR}{MetaboliteDatabase}(db, lower, upper)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{lower}{lower mass boundary}

\item{upper}{upper mass boundary}
}
\description{
Filter a MetaboliteDatabase for a given mass range.
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids))
db <- filterMR(db,100,120)
}
