% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalGroups.R
\name{functionalGroups}
\alias{functionalGroups}
\title{Extracting count data for functional groups}
\usage{
functionalGroups(x)
}
\arguments{
\item{x}{a character string of a valid SMILE}
}
\value{
a \code{tibble} of functional group counts
}
\description{
Extracting count data for functional groups
}
\examples{
functionalGroups('CN1C=C(N=C1)CC(C(=O)O)N')

}
