% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/get-methods.R
\name{getAccessions}
\alias{getAccessions}
\alias{getAccessions,MetaboliteDatabase-method}
\title{getAccessions-MetaboliteDatabase}
\usage{
getAccessions(db)

\S4method{getAccessions}{MetaboliteDatabase}(db)
}
\arguments{
\item{db}{MetaboliteDatabase}
}
\description{
return accession data table from MetaboliteDatabase object
}
