% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/get-methods.R
\name{getDescriptors}
\alias{getDescriptors}
\alias{getDescriptors,MetaboliteDatabase-method}
\title{Retrieve descriptors}
\usage{
getDescriptors(db)

\S4method{getDescriptors}{MetaboliteDatabase}(db)
}
\arguments{
\item{db}{MetaboliteDatabase}
}
\description{
return descriptor data table from MetaboliteDatabase object
}
