% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMF.R
\name{getMF}
\alias{getMF}
\title{Molecular formula generator}
\usage{
getMF(explan.mz, ppm, mode, MFfilter = T, lat = list(c(C = 0, iC = 0, H = 0,
  iH = 0, N = 0, iN = 0, O = 0, iO = 0, F = 0, Na = 0, Si = 0, P = 0, S = 0, Cl
  = 0, iCl = 0, Br = 0, iBr = 0, K = 0, iK = 0), c(C = 41, iC = 0, H = 72, iH =
  0, N = 15, iN = 0, O = 30, iO = 0, F = 0, Na = 0, Si = 0, P = 2, S = 2, Cl =
  0, iCl = 0, Br = 0, iBr = 0, K = 0, iK = 0)))
}
\arguments{
\item{explan.mz}{}

\item{ppm}{}

\item{mode}{}

\item{MFfilter}{}

\item{lat}{}
}
\description{
Molecular formula generator
}

