% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{goldenRules}
\alias{goldenRules}
\title{Golden rule tests for molecular formulas}
\usage{
goldenRules(MF)
}
\arguments{
\item{MF}{a vector of molecular formulas}
}
\value{
A tibble containing golden rule heuristic check results. \code{TRUE} is returned where relevant the Golden Rule
is satisfied and \code{FALSE} where not. \code{NA} is returned where the rule is not relevant for a molecular formula.
}
\description{
Heuristic tests for molecular formulas based on the Golden Rules 2, 4, 5 and 6 from Kind et al 2007.
}
\examples{
goldenRules(c('H2O','C12H22O11'))
}
\references{
Kind, T. and Fiehn, O., 2007. Seven Golden Rules for heuristic filtering of molecular formulas obtained by
accurate mass spectrometry. \emph{BMC bioinformatics}, \emph{8}(1), pp.1-20.
}
