% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{goldenRulesScore}
\alias{goldenRulesScore}
\title{Molecular formula plausibility scores}
\usage{
goldenRulesScore(golden_rules)
}
\arguments{
\item{golden_rules}{a tibble containing golden rule heuristic checks results as from \code{goldenRules()}}
}
\value{
A tibble containing golden rules plausibility scores.
}
\description{
Percentage plausibility scores based on rules 2, 4, 5 and 6 Kind et al 2007.
}
