% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{heteroatomProportion}
\alias{heteroatomProportion}
\title{The proportion of hetoratoms in molecular formulas}
\usage{
heteroatomProportion(element_frequencies, heteroatoms = c("C", "H", "O"))
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}

\item{heteroatoms}{a vector of elements to count as heteroatoms. Oxygen is included a default as this is often advantageous for molecular formulas of biological origin.}
}
\value{
A tibble contianing the
}
\description{
Calculate the proportions of heteroatoms in molecular formulas.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  heteroatomProportion()
}
