% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{heteroatomRatioCheck}
\alias{heteroatomRatioCheck}
\title{Heteroatom ratio check}
\usage{
heteroatomRatioCheck(
  element_ratios,
  range = c("common", "extended", "extreme")
)
}
\arguments{
\item{element_ratios}{a tibble containing molecular formula elemental ratios}

\item{range}{the ratio threshold ranges as defined by rule #5 of the Seven Golden Rules by Kind et al 2007}
}
\value{
A tibble containing results of the heteroatom ratio checks.
}
\description{
Heteroatom ratio checks based on rule #5 of the Seven Golden Rules by Kind et al 2007.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  elementRatios() \%>\% 
  heteroatomRatioCheck()
}
