% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionisationProducts.R
\name{ionisationProducts}
\alias{ionisationProducts}
\title{Ionisation Products}
\usage{
ionisationProducts(SMILES, adductTable = adducts())
}
\arguments{
\item{SMILES}{a valid SMILES string}

\item{adductTable}{table of adduct rules. Defaults to adducts()}
}
\description{
Calculate ionisation products for a given SMILES.
}
\examples{
ionisationProducts(aminoAcids$SMILES[1])
}
