% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{ipMF}
\alias{ipMF}
\title{Ionisation product molecular formula generation}
\usage{
ipMF(
  mz,
  adduct = "[M+H]1+",
  isotope = NA,
  ppm = 5,
  LEWIS = TRUE,
  SENIOR = TRUE,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)
}
\arguments{
\item{mz}{Accurate \emph{m/z}}

\item{adduct}{ionisation product adduct}

\item{isotope}{ionisation product isotope}

\item{ppm}{ppm error tolerance threshold}

\item{LEWIS}{return only molecular formulas that pass the LEWIS check}

\item{SENIOR}{return only molecular formulas that pass the SENiOR check}

\item{adduct_rules_table}{tibble containing available adduct formation rules. Defaults to \code{adduct_rules()}.}

\item{isotope_rules_table}{tibble containing available isotopic rules. Defaults to \code{isotope_rules()}.}
}
\description{
Generate molecular formulas for a given ionisation product accurate \emph{m/z}.
}
\examples{
ipMF(118.08626,adduct = '[M+H]1+')
}
