% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoDistr.R
\name{isoDistr}
\alias{isoDistr}
\title{Isotope Distribution Calculater}
\usage{
isoDistr(mf, chrg = 0)
}
\arguments{
\item{mf}{the molecular formular to generate isotope distributions}

\item{chrg}{the charge to apply to isotope distribution generation}
}
\description{
Isotope Distribution Calculater
}
\examples{
res <- isoDistr('C4H5O5',chrg=-1)
}
\author{
Manfred Beckmann
}

