% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_distributions.R
\name{isotopeDistribution}
\alias{isotopeDistribution}
\title{Isotopic distribution calculator}
\usage{
isotopeDistribution(MF, charge, limit = 9e-05, element_table = elements())
}
\arguments{
\item{MF}{a molecular formula}

\item{charge}{total charge}

\item{limit}{the minimum relative abundance threshold}

\item{element_table}{a table containing the elemental information. Defaults to \code{elements()}.}
}
\value{
A tibble containing the isotope distribution for the molecular formula.
}
\description{
Isotopic distribution calculator
}
\examples{
isotopeDistribution(
  'C4H5O5',
  charge = -1)
}
