% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{isotopePossible}
\alias{isotopePossible}
\title{Molecular formula isotopic possiblity}
\usage{
isotopePossible(MF, isotope = "13C", isotope_rules_table = isotope_rules())
}
\arguments{
\item{MF}{a character vector of molecular formulas}

\item{isotope}{the isotope to check}

\item{isotope_rules_table}{tibble containing available isotopic rules. Defaults to \code{isotope_rules()}.}
}
\value{
A boolean vector specifying if the specified isotope is possible for the molecular formulas.
}
\description{
Check if an isotope is possible for a vector of molecular formulas.
}
\examples{
isotopePossible(c('C12H22O11','H2O'))
}
