% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotopes.R
\docType{data}
\name{Isotopes}
\alias{Isotopes}
\title{Isotopes}
\format{A tibble containing 8 rows and 3 columns.
\describe{
\item{Isotope}{Isotope name}
\item{Mass Difference}{Isotopic mass difference}
\item{Rule}{Elemental composition rules}
}}
\usage{
Isotopes
}
\description{
Isotope rules
}
\keyword{datasets}
