% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{lewis}
\alias{lewis}
\alias{senior}
\title{LEWIS and SENIOR checks}
\usage{
lewis(MF, valences = list(C = 4, H = 1, N = 3, O = 2, P = 3, S = 4))

senior(MF, valences = list(C = 4, H = 1, N = 3, O = 2, P = 3, S = 4))
}
\arguments{
\item{MF}{molecular formula}

\item{valences}{named list of element valences}
}
\value{
Boolean vector of check results for each molecular formula.
}
\description{
LEWIS molecular formula valence test and SENIOR test for the existence of molecular graphs.
}
\examples{
lewis(c('C12H22O11','C12H22NO11'))
senior(c('C12H22O11','C12H22NO11'))
}
