% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteDatabase.R
\name{metaboliteDatabase}
\alias{metaboliteDatabase}
\title{metaboliteDatabase}
\usage{
metaboliteDatabase(accessions, descriptors, connection = NULL,
  type = "local")
}
\description{
Build a metabolite database ready for use.
}
\examples{
db <- metaboliteDatabase(aminoAcids,descriptors(aminoAcids$SMILE))
}
