% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{metaboliteDB}
\alias{metaboliteDB}
\title{Create a metabolite database}
\usage{
metaboliteDB(entries, descriptors = chemicalDescriptors(entries))
}
\arguments{
\item{entries}{tibble containing accession information. If \code{type = 'remote'} this should be the name of the table containing the accession information within the SQL database.}

\item{descriptors}{tibble containing descriptor information as returned by \code{descriptors()}. If \code{type = 'remote'} this should be the name of the table containing the descriptor information within the SQL database.}
}
\description{
Build a metabolite database ready for use.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
}
