% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{MFscore}
\alias{MFscore}
\title{Molecular formula scoring}
\usage{
MFscore(mf)
}
\arguments{
\item{mf}{Molecular formula}
}
\description{
Molecular formula scoring based on elemental ratios
}
\examples{
MFscore('C12H22O11')
}
