% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPsearch.R
\name{PIPsearch}
\alias{PIPsearch}
\title{Putative Ionisation Product searching}
\usage{
PIPsearch(mz, DB, ppm = 5, mode = "n", add = NA, iso = NA,
  adducts = mzAnnotation::Adducts, isotopes = mzAnnotation::Isotopes)
}
\arguments{
\item{mz}{the accurate m/z to search}

\item{DB}{object of class \code{MetaboliteDatabase}.}

\item{ppm}{the parts per million threshold to search}

\item{mode}{either 'p' or 'n' specifiying the acquisition mode. Can be overrided using \code{add} to do mixed mode searches.}

\item{add}{a character \code{vector} specifying the adducts to search. If \code{NA}, all adducts for that acquisition mode will be used.}

\item{iso}{a character \code{vector} specifying the isotopes to search. If \code{NA} isotopes will not be searched.}

\item{adducts}{adduct table containing available adduct rules. Defaults to table returned by \code{availableAdducts()}.}

\item{isotopes}{isotope table containing available isotope rules. Defaults to table returned by \code{availableIsotopes()}.}
}
\description{
Putative Ionisation Product searching
}
\examples{
#res <- PIPsearch(133.01378,5)
}
\author{
Jasen Finch
}
