% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmError.R
\name{ppmError}
\alias{ppmError}
\title{ppmError}
\usage{
ppmError(measured, theoretical)
}
\arguments{
\item{measured}{measured m/z}

\item{theoretical}{theoretical m/z}
}
\description{
Calculate ppmError between a measured and theoretical m/z
}
\examples{
ppmError(118.08626,118.08647)
}
