% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{rdbe}
\alias{rdbe}
\title{Calculate the rings plus double bonds equivalent}
\usage{
rdbe(MF, valences = list(C = 4, H = 1, N = 3, O = 2, P = 3, S = 4))
}
\arguments{
\item{MF}{vector of molecular formula}

\item{valences}{named list of element valences}
}
\value{
A vector of RDBE values.
}
\description{
Calculate the rings plus double bonds equivalent (RDBE) for molecular formulas.
}
\examples{
rdbe(c('C12H22O11','C12H22NO11'))
}
