% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{count.elements}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CHNOSZ}{\code{\link[CHNOSZ:makeup]{count.elements}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

