% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationshipPredictor.R
\name{relationshipPredictor}
\alias{relationshipPredictor}
\title{m/z relationship prediction}
\usage{
relationshipPredictor(mz, mode, limit = 0.001, add = NULL)
}
\arguments{
\item{mz}{numeric \code{vector} of accurate m/z}

\item{mode}{string of either 'p' or 'n; denoting the acquisition mode}

\item{limit}{limit of deviation for thresholding associations. Defaults to 0.001}

\item{add}{character vector of adducts to use. If \code{NULL} all available adducts will be used}
}
\description{
adduct, isotope and biotransfromation prediction.
}
\examples{
res <- relationshipPredictor(c(132.03023,133.01425,133.03359,168.00691),'n')
}
\author{
Jasen Finch
}
