% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeConvert.R
\name{safeConvert}
\alias{safeConvert}
\title{Safe Convert}
\usage{
safeConvert(input, inputType, outputType)
}
\arguments{
\item{input}{a valid SMILE or Inchi}

\item{inputType}{either "smiles" or "inchi", denoting the input type}

\item{outputType}{either "smiles", "inchi" or "inchikey", denoting the output type}
}
\description{
safely convert between SMILES and Inchi and to InchiKey using callr to catch potential segfaults
}
\examples{
safeConvert(aminoAcids$SMILE[1],'smiles','inchi')
}
