% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMZedDB.R
\name{shinyMZedDB}
\alias{shinyMZedDB}
\title{shinyMZedDB}
\usage{
shinyMZedDB()
}
\description{
A shiny application for the putative annotation of high resolution m/z.
}
\examples{
\dontrun{
shinyMZedDB()
}
}
\author{
Jasen Finch
}
