% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartSearch.R
\name{smartsSearch}
\alias{smartsSearch}
\title{smartsSearch}
\usage{
smartsSearch(smiles, smarts)
}
\arguments{
\item{smiles}{a valid SMILES}

\item{smarts}{a valid SMARTS symbol}
}
\description{
SMARTS substructure searching for SMILES.
}
\examples{
smartsSearch(aminoAcids$SMILES[1],"[OX2H]")
}
