% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smilesToAccurateMass}
\alias{smilesToAccurateMass}
\title{Convert SMILES to accurate mass}
\usage{
smilesToAccurateMass(smile)
}
\arguments{
\item{smile}{a valid SMILE}
}
\description{
convert a smile to an accurate mass
}
\examples{
smilesToAccurateMass(amino_acids$SMILES[1])
}
