% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smilesToAccurateMasses.R
\name{smilesToAccurateMasses}
\alias{smilesToAccurateMasses}
\title{Convert SMILES to Accurate Masses}
\usage{
smilesToAccurateMasses(smiles)
}
\arguments{
\item{smile}{a character string of a valid SMILE}
}
\value{
a \code{tibble} of Accurate Masses
}
\description{
Convert SMILES to Accurate Masses
}
\examples{
smilesToAccurateMasses('C[C@H](C(=O)O)N')
}
