% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smilesToMF}
\alias{smilesToMF}
\title{Convert SMILES to molecular formula}
\usage{
smilesToMF(smile)
}
\arguments{
\item{smile}{a valid SMILE}
}
\description{
convert a smile to a molecular formula
}
\examples{
smilesToMF(amino_acids$SMILES[1])
}
