% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smilesToMFs.R
\name{smilesToMFs}
\alias{smilesToMFs}
\title{Convert SMILES to Molecular Formulas}
\usage{
smilesToMFs(smiles)
}
\arguments{
\item{smile}{a character string of a valid SMILE}
}
\value{
a \code{tibble} of Molecular Formulas
}
\description{
Convert SMILES to Molecular Formulas
}
\examples{
smilesToMFs('C[C@H](C(=O)O)N')
}
