% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smileToAccurateMass}
\alias{smileToAccurateMass}
\title{Convert SMILES to accurate mass}
\usage{
smileToAccurateMass(smile)
}
\arguments{
\item{smile}{a valid SMILE}
}
\description{
convert a smile to an accurate mass
}
\examples{
smileToAccurateMass(amino_acids$SMILES[1])
}
