% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{suitableElementRanges}
\alias{suitableElementRanges}
\title{Calculate suitable elemental frequency ranges}
\usage{
suitableElementRanges(mass)
}
\arguments{
\item{mass}{molecular mass}
}
\value{
A list of minimum and maximum frequencies for each element.
}
\description{
Calculate elemental frequency ranges for a given mass which are suitable for molecular formula generation.
}
\examples{
suitableElementRanges(342.11621)
}
