% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMF.R
\name{suitableMaxElements}
\alias{suitableMaxElements}
\title{Calculate suitable maximum elemental frequencies}
\usage{
suitableMaxElements(mass)
}
\arguments{
\item{mass}{accurate mass}
}
\value{
named numeric vector of element frequencies
}
\description{
Calculate maximum elemental frequencies for a given mass which are suitable for molecular formula generation.
}
