% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{transformationPossible}
\alias{transformationPossible}
\title{Transformation check}
\usage{
transformationPossible(
  from,
  to,
  transformation,
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{from}{molecular formula to be transformed}

\item{to}{molecular formula product of the transformation}

\item{transformation}{transformation to apply. As found in column \verb{MF Change} of the table supplied to the arguement \code{transformation_rules_table}.}

\item{transformation_rules_table}{the transformation rules table. Defaults to the returned value of \code{transformation_rules()}. Alternative tables should be supplied in the same format.}
}
\value{
TRUE/FALSE depending whether the transformation is possible
}
\description{
Check if a transformation between two molecular formulas is possible.
}
\examples{
transformationPossible(
  'C12H22O11',
  'C6H12O6',
  'M + [H2]'
  )
}
