% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{transformation_rules}
\alias{transformation_rules}
\alias{transformation_names}
\title{Transformation rules}
\format{
A tibble containing 12 rows and 9 columns.
\describe{
\item{Name}{Transformation name}
\item{MF Change}{Molecular formula change}
\item{Difference}{Mass difference}
\item{C}{Change in carbon atoms}
\item{H}{Change in hydrogen atoms}
\item{O}{Change in oxygen atoms}
\item{N}{Change in nitrogen atoms}
\item{P}{Change in phosphorus atoms}
\item{S}{Change in sulphur atoms}
}
}
\usage{
transformation_rules()

transformation_names()
}
\description{
Transformation rules
}
