% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformMF.R
\name{transformMF}
\alias{transformMF}
\title{transformMF}
\usage{
transformMF(MF, transformation = "M - [O] + NH2]",
  transformations = mzAnnotation::Transformations)
}
\arguments{
\item{MF}{molecular formula to transform}

\item{transformation}{transformation to apply}

\item{transformations}{transformations table containing available transformations rules. Defaults to \code{Transformations}.}
}
\description{
transform a molecular formula
}
\details{
\code{NA} will be returned if \code{MF} cannot be transformed.
}
\examples{
transformMF('C4H5O5')
transformMF('C4H5N',transformation = 'M - [OH] + [NH2]')
}
