% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{filterMR}
\alias{filterMR}
\alias{filterMR,MetaboliteDatabase-method}
\alias{filterER}
\alias{filterER,MetaboliteDatabase-method}
\alias{filterIP}
\alias{filterIP,MetaboliteDatabase-method}
\alias{filterACCESSIONS}
\alias{filterACCESSIONS,MetaboliteDatabase-method}
\alias{filterMF}
\alias{filterMF,MetaboliteDatabase-method}
\alias{calcAdducts}
\alias{calcAdducts,MetaboliteDatabase-method}
\alias{PIPsearch}
\alias{PIPsearch,MetaboliteDatabase-method}
\alias{elementFreq}
\alias{elementFreq,MetaboliteDatabase-method}
\title{Filter a mass range}
\usage{
filterMR(db, lower, upper)

\S4method{filterMR}{MetaboliteDatabase}(db, lower, upper)

filterER(db, rule)

\S4method{filterER}{MetaboliteDatabase}(db, rule)

filterIP(db, rule)

\S4method{filterIP}{MetaboliteDatabase}(db, rule)

filterACCESSIONS(db, ids)

\S4method{filterACCESSIONS}{MetaboliteDatabase}(db, ids)

filterMF(db, mf)

\S4method{filterMF}{MetaboliteDatabase}(db, mf)

calcAdducts(db, id, adduct_rules_table = adduct_rules())

\S4method{calcAdducts}{MetaboliteDatabase}(db, id, adduct_rules_table = adduct_rules())

PIPsearch(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)

\S4method{PIPsearch}{MetaboliteDatabase}(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)

elementFreq(db)

\S4method{elementFreq}{MetaboliteDatabase}(db)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{lower}{lower mass boundary}

\item{upper}{upper mass boundary}
}
\description{
Filter a MetaboliteDatabase for a given mass range.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
db <- filterMR(db,100,120)
}
