% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{entries}
\alias{entries}
\alias{entries,MetaboliteDatabase-method}
\alias{descriptors}
\alias{descriptors,MetaboliteDatabase-method}
\title{Retrieve database entries}
\usage{
entries(db)

\S4method{entries}{MetaboliteDatabase}(db)

descriptors(db)

\S4method{descriptors}{MetaboliteDatabase}(db)
}
\arguments{
\item{db}{object of S4 class \code{MetaboliteDatabase}}
}
\description{
Get or set tables in a \code{MetaboliteDatabase} class object.
}
