% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adducts.R
\docType{data}
\name{Adducts}
\alias{Adducts}
\title{Adducts}
\format{A tibble containing 58 rows and 6 columns
\describe{
\item{Name}{Adduct name}
\item{Charge}{Charge added}
\item{xM}{Number of M needed}
\item{Add}{Mass value change}
\item{Nelec}{Electron change}
\item{Rule}{Structural rule for formation}
}}
\usage{
Adducts
}
\description{
Formation rules for electrospray ionisation adducts
}
\keyword{datasets}
