% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductTransfromMF.R
\name{adductTransformMF}
\alias{adductTransformMF}
\title{adductTransfromMF}
\usage{
adductTransformMF(MF, adduct)
}
\arguments{
\item{MF}{molecular formula to transform}

\item{adduct}{adduct to use for transformation}
}
\description{
adduct transform a molecular formula
}
\examples{
adductTransfromMF('C6H12O6','[M+H]1+')
}
