% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductTransfromMF.R
\name{adductTransformMF}
\alias{adductTransformMF}
\title{Molecular formula adduct transformation}
\usage{
adductTransformMF(MF, adduct, Adducts = adducts())
}
\arguments{
\item{MF}{molecular formula to transform.}

\item{adduct}{adduct to use for transformation.}

\item{Adducts}{Adduct information table to use. Defaults to \code{mzAnnotation::Adducts}.}
}
\description{
adduct transform a molecular formula.
}
\examples{
adductTransformMF('C6H12O6','[M+H]1+')
}
