% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoAcids.R
\docType{data}
\name{aminoAcids}
\alias{aminoAcids}
\title{aminoAcids}
\format{
a tibble containing 20 rows and 6 columns
\describe{
\item{ACCESSION_ID}{}
\item{NAME}{}
\item{InChi}{}
\item{InChiKey}{}
\item{SMILES}{}
}
}
\usage{
aminoAcids
}
\description{
Chemical structures of the essential amino acids
}
\keyword{datasets}
