% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcAccurateMass}
\alias{calcAccurateMass}
\title{Calculate a molecular formula  accurate mass}
\usage{
calcAccurateMass(MF, charge = 0, element_table = elements())
}
\arguments{
\item{MF}{the molecular formula for which to calculate the accurate mass}

\item{charge}{the total charge of the specified molecular formula}

\item{element_table}{elemental mass information. Defaults to \code{\link{elements}()}.}
}
\description{
Calculate the accurate mass of a specified molecular formula.
}
\examples{
calcAccurateMass('C4H5O5',charge = 0)
}
