% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAccurateMass.R
\name{calcAccurateMass}
\alias{calcAccurateMass}
\title{calcAccurateMass}
\usage{
calcAccurateMass(MF, charge = 0, elementTable = elements())
}
\arguments{
\item{MF}{molecular formula for which to calculate the accrate mass}

\item{charge}{charge of the given molecular formula}

\item{elementTable}{element information table. Defaults to \code{\link{elements}()}.}
}
\description{
calculate the accurate mass of a given molecular formula
}
\examples{
calcAccurateMass('C4H5O5',charge = 0)
}
