% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/calcAdducts.R
\name{calcAdducts}
\alias{calcAdducts}
\title{calcAdducts}
\usage{
calcAdducts(db, id, adductTable = adducts())

\S4method{calcAdducts}{MetaboliteDatabase}(db, id, adductTable = adducts())
}
\arguments{
\item{id}{accession id}

\item{adductTable}{table of adduct rules. Defaults to adducts()}
}
\description{
Calculate ionisation products for given database accession.
#' @param db object of class \code{MetaboliteDatabase}
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids$SMILES))
add <- calcAdducts(db,1)
}
