% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{calcAdducts}
\alias{calcAdducts}
\alias{calcAdducts,MetaboliteDatabase-method}
\title{Calculate adducts}
\usage{
calcAdducts(db, id, adduct_rules_table = adduct_rules())

\S4method{calcAdducts}{MetaboliteDatabase}(db, id, adduct_rules_table = adduct_rules())
}
\arguments{
\item{db}{object of class \code{MetaboliteDatabase}}

\item{id}{accession id}

\item{adduct_rules_table}{table of adduct rules. Defaults to adducts()}
}
\description{
Calculate ionisation products for given database accession.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
add <- calcAdducts(db,1)
}
