% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcDiff.R
\name{calcDiff}
\alias{calcDiff}
\title{Identify structural relationships based on m/z differences}
\usage{
calcDiff(mat, mo)
}
\arguments{
\item{mat}{}

\item{mo}{}
}
\description{
Identify structural relationships based on m/z differences
}

