% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcM.R
\name{calcM}
\alias{calcM}
\title{calcM}
\usage{
calcM(mz, adduct = "[M+H]1+", isotope = NA, transformation = NA,
  adducts = mzAnnotation::Adducts, isotopes = mzAnnotation::Isotopes,
  transformations = mzAnnotation::Transformations)
}
\arguments{
\item{mz}{m/z for which to calculate M}

\item{adduct}{adduct to apply}

\item{isotope}{isotope to apply}

\item{transformation}{transformation to apply}

\item{adducts}{adduct table containing available adduct rules. Defaults to table returned by \code{availableAdducts()}.}

\item{isotopes}{isotope table containing available isotope rules. Defaults to table returned by \code{availableIsotopes()}.}

\item{transformations}{transformations table containing available transformations rules. Defaults to table returned by \code{availableTransformations()}.}
}
\description{
calculate M for a given m/z, adduct, isotope and transformation
}
\examples{
calcM(118.08626,adduct = '[M+H]1+',isotope = 'C13',transformation = 'M - [O] + NH2]')
}
