% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcM.R
\name{calcM}
\alias{calcM}
\title{calcM}
\usage{
calcM(mz, adduct = "[M+H]1+", isotope = NA, transformation = NA,
  adductTable = adducts(), isotopeTable = isotopes(),
  transformationTable = transformations())
}
\arguments{
\item{mz}{m/z for which to calculate M}

\item{adduct}{adduct to apply}

\item{isotope}{isotope to apply}

\item{transformation}{transformation to apply}

\item{adductTable}{adduct table containing available adduct rules. Defaults to \code{Adducts}.}

\item{isotopeTable}{isotope table containing available isotope rules. Defaults to \code{Isotopes}.}

\item{transformationTable}{transformations table containing available transformations rules. Defaults to \code{Transformations}.}
}
\description{
calculate M for a given m/z, adduct, isotope and transformation
}
\examples{
calcM(118.08626,adduct = '[M+H]1+',isotope = '13C',transformation = 'M - [O] + NH2]')
}
