% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcMZ}
\alias{calcMZ}
\title{Calculate the m/z of a mass}
\usage{
calcMZ(
  M,
  adduct = "[M+H]1+",
  isotope = NA,
  transformation = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules(),
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{M}{M for which to calculate an m/z}

\item{adduct}{adduct to apply}

\item{isotope}{isotope to apply}

\item{transformation}{transformation to apply}

\item{adduct_rules_table}{adduct table containing available adduct rules. Defaults to \code{adducts()}.}

\item{isotope_rules_table}{isotope table containing available isotope rules. Defaults to \code{isotopes()}.}

\item{transformation_rules_table}{transformations table containing available transformations rules. Defaults to \code{transformations()}.}
}
\description{
calculate an m/z for a given M, adduct, isotope and transformation
}
\examples{
calcMZ(116.05182,adduct = '[M+H]1+',isotope = '13C',transformation = 'M - [O] + NH2]')
}
