% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{CHOproportion}
\alias{CHOproportion}
\title{The proportion of C, H and O atoms in molecular formulas}
\usage{
CHOproportion(element_frequencies)
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble containing the CHO proportions for the specified element frequencies.
}
\description{
Calculate the proportion of C, H and O in specified molecular formulas.
}
\examples{
c('H2O','C12H22O11') \%>\% 
  elementFrequencies() \%>\% 
  CHOproportion()
}
