% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert chemical notation}
\usage{
convert(input, inputType, outputType)
}
\arguments{
\item{input}{a valid SMILE or Inchi}

\item{inputType}{either "smiles" or "inchi", denoting the input type}

\item{outputType}{either "smiles", "inchi" or "inchikey", denoting the output type}
}
\description{
convert between SMILES and Inchi and to InchiKey
}
\examples{
convert(amino_acids$SMILES[1],'smiles','inchi')
}
