% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corAnalysis.R
\name{corAnalysis}
\alias{corAnalysis}
\title{Correlation Analysis for mz data matirx}
\usage{
corAnalysis(data,varlist,pval=0.05)
}
\arguments{
\item{data}{}

\item{varlist}{}

\item{pval}{}
}
\description{
Calculates correlations with a p-value above a given threshold between all variables in supplied data. M/z relationships between correlated vairables are given for a supplied variable list.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
\seealso{
corLists, calcDiff
}

