% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{descriptors}
\alias{descriptors}
\title{convert SMILES to a series of molecular descriptors}
\usage{
descriptors(smiles)
}
\arguments{
\item{smiles}{vector of valid SMILEs}
}
\description{
convert SMILES to a series of molecular descriptors
}
\examples{
data(aminoAcids)
descriptors(aminoAcids$SMILE)
}
