% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{descriptors}
\alias{descriptors}
\title{Convert SMILE to a series of molecular descriptors}
\usage{
descriptors(smile)
}
\arguments{
\item{smile}{a valid SMILE}
}
\value{
a \code{tibble} of the following;
\itemize{
    \item{HBD} The number of Hydrogen Bond Donors
    \item{HBA} The number of Hydrogen Bond Acceptors
    \item{AcidGroups} The number of Acidic Groups
    \item{BaseGroups} The number of Basic Groups
    \item{TotalCharge} The total overall charge
}
}
\description{
Convert SMILE to a series of molecular descriptors
}
\examples{
descriptors('CN1C=C(N=C1)CC(C(=O)O)N')
}
