% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{elementCountCheck}
\alias{elementCountCheck}
\title{Element count checks}
\usage{
elementCountCheck(element_frequencies)
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble containing results of the element count checks.
}
\description{
Element count checks based on rule #6 of the Seven Golden Rules by Kind et al 2007.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  elementCountCheck()
}
