% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{elementRatios}
\alias{elementRatios}
\title{Calculate element ratios}
\usage{
elementRatios(element_frequencies, elements = c("C", "H", "N", "O", "P", "S"))
}
\arguments{
\item{element_frequencies}{a tibble containing element frequencies as returned by \code{elementFrequencies()}}
}
\value{
A tibble containing element frequency ratios.
}
\description{
Calculate ratios between all element frequency combinations.
}
\examples{
elementFrequencies(c('H2O','C12H22O11')) \%>\% 
  elementRatios()
}
