% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metabolite_database.R
\name{filterIP}
\alias{filterIP}
\alias{filterIP,MetaboliteDatabase-method}
\title{Filter using an ionisation rule}
\usage{
filterIP(db, rule)

\S4method{filterIP}{MetaboliteDatabase}(db, rule)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{rule}{Character containing ionisation rule.}
}
\description{
Filter MetaboliteDatabase based on an ionisation rule
}
\examples{
rule <- adduct_rules()$Rule[52]
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
db <- filterIP(db,rule)
}
