% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/filter-methods.R
\docType{methods}
\name{filterMF}
\alias{filterMF}
\alias{filterMF,MetaboliteDatabase-method}
\title{filterMF}
\usage{
filterMF(db, mf)

\S4method{filterMF}{MetaboliteDatabase}(db, mf)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{mf}{character vector of molecular formulas}
}
\description{
Filter a MetaboliteDatabase based on given molecular formulas
}
\examples{
db <- metaboliteDB(aminoAcids,descriptors(aminoAcids))
db <- filterMF(db,c('C3H7NO2','C5H10N2O3'))
}
