% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metabolite_database.R
\name{filterMR}
\alias{filterMR}
\alias{filterMR,MetaboliteDatabase-method}
\title{Filter a mass range}
\usage{
filterMR(db, lower, upper)

\S4method{filterMR}{MetaboliteDatabase}(db, lower, upper)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{lower}{lower mass boundary}

\item{upper}{upper mass boundary}
}
\description{
Filter a MetaboliteDatabase for a given mass range.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
db <- filterMR(db,100,120)
}
