% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalGroups.R
\name{FunctionalGroups}
\alias{FunctionalGroups}
\title{Extracting count data for functional groups}
\usage{
FunctionalGroups(x)
}
\arguments{
\item{x}{a character string of a valid SMILE}
}
\value{
a \code{tibble} of functional group counts
}
\description{
Extracting count data for functional groups
}
\examples{
FunctionalGroups('CN1C=C(N=C1)CC(C(=O)O)N')

}
