% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMF.R
\name{generateMF}
\alias{generateMF}
\title{generateMF}
\usage{
generateMF(mass, ppm = 1, charge = 0, validation = TRUE,
  element_max = c(C = 12, H = 22, N = 0, O = 11, P = 0, S = 0),
  element_min = c(C = 0, H = 0, N = 0, O = 0, P = 0, S = 0))
}
\arguments{
\item{mass}{accurate mass}

\item{ppm}{ppm tolerance}

\item{charge}{charge}

\item{validation}{\code{boolean}, apply validation rules}

\item{element_max}{numeric \code{vector} of maximum elemental composition}

\item{element_min}{numeric \code{vector} of minimum elemental composition}
}
\value{
A \code{tibble} containing the generated MFs, their theoretical mass and their PPM error.
}
\description{
Molecular formula generation
}
\details{
this uses the HR2 molecular formula generator available at \url{http://maltese.dbs.aber.ac.uk:8888/hrmet/supp/rhrmet.html}.
}
\examples{
res <- generateMF(342.11621,
                  element_max = c(C = 12,H = 22,N = 0,
                                O = 11,P = 0,S = 0))
}
\author{
Jasen Finch
}
