% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMF.R
\name{generateMF}
\alias{generateMF}
\title{Molecular formula generation}
\usage{
generateMF(mass, ppm = 1, charge = 0, validation = TRUE,
  composition = c(C = 12, H = 22, N = 0, O = 11, P = 0, S = 0),
  generator = "HR2")
}
\arguments{
\item{mass}{accurate mass for MF generation}

\item{ppm}{ppm tolerance for MF generation}

\item{charge}{charge to apply to MF generation}

\item{validation}{\code{boolean} apply validation rules}

\item{composition}{numeric \code{vector} of maximum elemental composition}
}
\value{
A \code{data.frame} containing the generated MFs, their theoretical mass and their PPM error.
}
\description{
Molecular formula generation
}
\details{
Uses the HR2 molecular formula generator available at \url{http://maltese.dbs.aber.ac.uk:8888/hrmet/supp/rhrmet.html}.
}
\examples{
res <- generateMF(342.11621,
                  composition=c(C = 12,H = 22,N = 0,
                                O = 11,P = 0,S = 0))
}
\author{
Jasen Finch
}
