% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{generateMF}
\alias{generateMF}
\title{Molecular formula generation}
\usage{
generateMF(
  mass,
  ppm = 1,
  charge = 0,
  element_ranges = suitableElementRanges(mass),
  LEWIS = TRUE,
  SENIOR = TRUE
)
}
\arguments{
\item{mass}{accurate mass}

\item{ppm}{ppm tolerance}

\item{charge}{charge}

\item{element_ranges}{named list of element}

\item{LEWIS}{return only molecular formulas that pass the LEWIS check}

\item{SENIOR}{return only molecular formulas that pass the SENiOR check}
}
\value{
A \code{tibble} containing the generated MFs, their theoretical mass and their PPM error.
}
\description{
Molecular formula generation
}
\examples{
generateMF(342.11621,
           element_ranges = list(C = c(0,12),
                                 H = c(0,22),
                                 O = c(0,11)))
}
\author{
Jasen Finch
}
