% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{goldenRules}
\alias{goldenRules}
\title{Golden rule tests for molecular formlas}
\usage{
goldenRules(MF)
}
\arguments{
\item{MF}{a vector of molecular formulas}
}
\value{
A tibble containing golden rule heuristic check results.
}
\description{
Heuristic tests for moleucla formulas based on the golden rules 2, 4, 5 and 6 from Kind et al 2007.
}
\examples{
goldenRules(c('H2O','C12H22O11'))
}
