% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionisationProducts.R
\name{ionisationProducts}
\alias{ionisationProducts}
\title{Ionisation products}
\usage{
ionisationProducts(SMILES, adduct_rules_table = adduct_rules())
}
\arguments{
\item{SMILES}{a valid SMILES string}

\item{adduct_rules_table}{table of adduct rules. Defaults to adducts()}
}
\description{
Calculate ionisation products for a given SMILES.
}
\examples{
ionisationProducts(amino_acids$SMILES[1])
}
