% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{ipMF}
\alias{ipMF}
\title{Ionisation product molecular formula generation}
\usage{
ipMF(
  mz,
  adduct = "[M+H]1+",
  isotope = NA,
  ppm = 5,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)
}
\arguments{
\item{mz}{accurate \emph{m/z}}

\item{adduct}{ionisation product adduct}

\item{isotope}{ionisation product isotope}

\item{ppm}{parts per million error tolerance threshold}

\item{adduct_rules_table}{a tibble containing available adduct formation rules. Defaults to \code{adduct_rules()}.}

\item{isotope_rules_table}{a tibble containing available isotopic rules. Defaults to \code{isotope_rules()}.}
}
\description{
Generate molecular formulas for a specified ionisation product accurate \emph{m/z}.
}
\examples{
ipMF(118.08626,adduct = '[M+H]1+')
}
