% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeDistribution.R
\name{isotopeDistribution}
\alias{isotopeDistribution}
\title{Isotope Distribution Calculator}
\usage{
isotopeDistribution(MF, charge, limit = 9e-05, elements = elements())
}
\arguments{
\item{MF}{the molecular formular to generate the isotope distribution}

\item{charge}{the charge of the molecular formula}

\item{limit}{the relative abundance threshold}

\item{elements}{the table containing elemental information. Defaults to \code{Elements}.}
}
\description{
Isotope Distribution Calculator
}
\examples{
res <- isotopeDistribution('C4H5O5',charge = -1)
}
\author{
Jasen Finch
}
