% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{isotope_rules}
\alias{isotope_rules}
\alias{isotope_names}
\title{Isotopic rules}
\format{
A tibble containing 8 rows and 3 columns.
\describe{
\item{Isotope}{Isotope name}
\item{Mass Difference}{Isotopic mass difference}
\item{Rule}{Elemental composition rules}
}
}
\usage{
isotope_rules()

isotope_names()
}
\description{
Isotope rules
}
