% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/PIPsearch.R
\name{PIPsearch}
\alias{PIPsearch}
\alias{PIPsearch,MetaboliteDatabase-method}
\title{Putative Ionisation Product searching}
\usage{
PIPsearch(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  isotopeTable = isotopes(),
  adductTable = adducts()
)

\S4method{PIPsearch}{MetaboliteDatabase}(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  isotopeTable = isotopes(),
  adductTable = adducts()
)
}
\arguments{
\item{db}{object of class \code{MetaboliteDatabase}.}

\item{mz}{the accurate m/z to search.}

\item{ppm}{the parts per million threshold to search.}

\item{adduct}{the adduct name to search.}

\item{isotope}{the isotope name to search. Defaults to NA for non-isotopic searches.}

\item{isotopeTable}{isotope table containing available isotope rules. Defaults to table returned by \code{isotopes()}.}

\item{adductTable}{adduct table containing available adduct rules. Defaults to table returned by \code{adducts()}.}
}
\description{
Putative Ionisation Product searching
}
\examples{
res <- PIPsearch(metaboliteDB(aminoAcids,descriptors(aminoAcids$SMILES)),132.03023,5,'[M-H]1-')
}
\author{
Jasen Finch
}
