% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPsearch.R
\name{PIPsearch}
\alias{PIPsearch}
\title{Putative Ionisation Product searching}
\usage{
PIPsearch(mz, db, ppm, adduct, isotope = NA,
  isotopes = mzAnnotation::Isotopes, adducts = mzAnnotation::Adducts)
}
\arguments{
\item{mz}{the accurate m/z to search}

\item{ppm}{the parts per million threshold to search}

\item{isotopes}{isotope table containing available isotope rules. Defaults to table returned by \code{availableIsotopes()}.}

\item{adducts}{adduct table containing available adduct rules. Defaults to table returned by \code{availableAdducts()}.}

\item{mode}{either 'p' or 'n' specifiying the acquisition mode. Can be overrided using \code{add} to do mixed mode searches.}

\item{add}{a character \code{vector} specifying the adducts to search. If \code{NA}, all adducts for that acquisition mode will be used.}

\item{iso}{a character \code{vector} specifying the isotopes to search. If \code{NA} isotopes will not be searched.}

\item{DB}{object of class \code{MetaboliteDatabase}.}
}
\description{
Putative Ionisation Product searching
}
\examples{
res <- PIPsearch(132.03023,metaboliteDatabase(aminoAcids,descriptors(aminoAcids)),5,'[M-H]1-')
}
\author{
Jasen Finch
}
