% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPsearch.R
\name{PIPsearch}
\alias{PIPsearch}
\title{Putative Ionisation Product searching}
\usage{
PIPsearch(mz, db, ppm, adduct, isotope = NA,
  isotopes = mzAnnotation::Isotopes, adducts = mzAnnotation::Adducts)
}
\arguments{
\item{mz}{the accurate m/z to search.}

\item{db}{object of class \code{MetaboliteDatabase}.}

\item{ppm}{the parts per million threshold to search.}

\item{adduct}{the adduct name to search.}

\item{isotope}{the isotope name to search. Defaults to NA for non-isotopic searches.}

\item{isotopes}{isotope table containing available isotope rules. Defaults to table returned by \code{availableIsotopes()}.}

\item{adducts}{adduct table containing available adduct rules. Defaults to table returned by \code{availableAdducts()}.}
}
\description{
Putative Ionisation Product searching
}
\examples{
res <- PIPsearch(132.03023,metaboliteDB(aminoAcids,descriptors(aminoAcids)),5,'[M-H]1-')
}
\author{
Jasen Finch
}
