% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPsearch.R
\name{PIPsearch}
\alias{PIPsearch}
\alias{PIPsearch,MetaboliteDatabase-method}
\title{Putative ionisation product search}
\usage{
PIPsearch(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)

\S4method{PIPsearch}{MetaboliteDatabase}(
  db,
  mz,
  ppm,
  adduct,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)
}
\arguments{
\item{db}{object of class \code{MetaboliteDatabase}.}

\item{mz}{the accurate m/z to search.}

\item{ppm}{the parts per million threshold to search.}

\item{adduct}{the adduct name to search.}

\item{isotope}{the isotope name to search. Defaults to NA for non-isotopic searches.}

\item{adduct_rules_table}{adduct formation rules tabl. Defaults to table returned by \code{adducts_rules()}.}

\item{isotope_rules_table}{isotope table containing available isotope rules. Defaults to table returned by \code{isotope_rules()}.}
}
\description{
Putative ionisation product search
}
\examples{
res <- PIPsearch(metaboliteDB(amino_acids,descriptors(amino_acids$SMILES)),132.03023,5,'[M-H]1-')
}
\author{
Jasen Finch
}
