% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ppmError}
\alias{ppmError}
\title{Calculate PPM error}
\usage{
ppmError(measured, theoretical)
}
\arguments{
\item{measured}{measured \emph{m/z}}

\item{theoretical}{theoretical \emph{m/z}}
}
\description{
Calculate the parts per million error between a measured and theoretical \emph{m/z}.
}
\examples{
ppmError(118.08626,118.08647)
}
