% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppmError}
\alias{ppmError}
\title{Calculate ppm error}
\usage{
ppmError(measured, theoretical)
}
\arguments{
\item{measured}{measured mass to charge ratio}

\item{theoretical}{theoretical mass to charge ratio}
}
\value{
The parts per million error.
}
\description{
Calculate the parts per million error between a measured \emph{m/z} and a theoretical \emph{m/z}.
}
\examples{
ppmError(118.08626,118.08647)
}
