% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ppmRange}
\alias{ppmRange}
\title{Calculate a PPM error range}
\usage{
ppmRange(mz, ppm)
}
\arguments{
\item{mz}{the m/z for which to calculate the range}

\item{ppm}{the ppm}
}
\description{
Calculate the upper and lower ppm boundaries for a given m/z
}
\examples{
ppmRange(118.08626,5)
}
