% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_rules.R
\name{rdbe}
\alias{rdbe}
\title{Calculate the rings plus double bonds equivalent}
\usage{
rdbe(
  element_frequencies,
  valences = list(C = 4, H = 1, N = 3, O = 2, P = 3, S = 4)
)
}
\arguments{
\item{element_frequencies}{table of element frequencies for a set of molecular formulas as returned by \code{elementFrequencies()}.}

\item{valences}{named list of element valences}
}
\value{
A vector of RDBE values.
}
\description{
Calculate the rings plus double bonds equivalent (RDBE) for molecular formulas.
}
\examples{
element_frequencies <- elementFrequencies(c('C12H22O11','C12H22NO11'))
rdbe(element_frequencies)
}
