% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_calculaton.R
\name{relationshipCalculator}
\alias{relationshipCalculator}
\title{m/z relationship calculation}
\usage{
relationshipCalculator(
  mz,
  limit = 0.001,
  adducts = c("[M-H]1-", "[M+Cl]1-", "[M+K-2H]1-"),
  isotopes = NA,
  transformations = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules(),
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{mz}{a vector of mass to charge ratios}

\item{limit}{the limit of mass deviation for associations. Defaults to 0.001}

\item{adducts}{character vector of possible adducts names to test}

\item{isotopes}{character vector of possible isotopes to test}

\item{transformations}{character vector of possible transformations to test}

\item{adduct_rules_table}{table containing adduct formation rules. Defaults to \code{adduct_rules()}.}

\item{isotope_rules_table}{table containing isotope rules. Defaults to \code{isotope_rules()}.}

\item{transformation_rules_table}{table containing transformation rules. Defaults to \code{transformation_rules()}.}
}
\value{
A tibble of possible adduct, isotope and transformation relationships between the specified \emph{m/z}.
}
\description{
Calculate possible adduct, isotope and transformation relationships between \emph{m/z}.
}
\examples{
relationshipCalculator(c(132.03023,168.00691))
}
