% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationshipCalculator.R
\name{relationshipCalculator}
\alias{relationshipCalculator}
\title{m/z relationship calculation}
\usage{
relationshipCalculator(
  mz,
  limit = 0.001,
  adducts = c("[M-H]1-", "[M+Cl]1-", "[M+K-2H]1-"),
  isotopes = NA,
  transformations = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules(),
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{mz}{numeric \code{vector} of accurate m/z}

\item{limit}{limit of deviation for thresholding associations. Defaults to 0.001}

\item{adducts}{character vector of adducts names to use}

\item{isotopes}{character vector of isotopes to use.}

\item{transformations}{character vector of transformations to use}

\item{adduct_rules_table}{table containing adduct formation rules. Defaults to \code{adduct_rules()}.}

\item{isotope_rules_table}{table containing isotope rules. Defaults to \code{isotope_rules()}.}

\item{transformation_rules_table}{table containing transformation rules. Defaults to \code{transformation_rules()}.}
}
\description{
adduct, isotope and biotransfromation calculation.
}
\examples{
relationshipCalculator(c(132.03023,168.00691))
}
\author{
Jasen Finch
}
