% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationshipPredictor.R
\name{relationshipPredictor}
\alias{relationshipPredictor}
\title{m/z relationship prediction}
\usage{
relationshipPredictor(mz, limit = 0.001, adducts = c("[M-H]1-", "[M+Cl]1-",
  "[M+K-2H]1-"), isotopes = NULL, transformations = NULL,
  adductTable = mzAnnotation::Adducts,
  isotopeTable = mzAnnotation::Isotopes,
  transformationTable = mzAnnotation::Transformations)
}
\arguments{
\item{mz}{numeric \code{vector} of accurate m/z}

\item{limit}{limit of deviation for thresholding associations. Defaults to 0.001}

\item{adducts}{character vector of adducts to use. If \code{NULL} all available adducts will be used.}

\item{isotopes}{character vector of isotopes to use.}

\item{transformations}{character vector of transformations to use}
}
\description{
adduct, isotope and biotransfromation prediction.
}
\examples{
relationshipPredictor(c(132.03023,168.00691))
}
\author{
Jasen Finch
}
