% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationshipPredictor.R
\name{relationshipPredictor}
\alias{relationshipPredictor}
\title{adduct relationship prediction}
\usage{
relationshipPredictor(mz, mode, limit = 0.001)
}
\arguments{
\item{mz}{numeric \code{vector} of accurate m/z}

\item{mode}{string of either 'p' or 'n; denoting the acquisition mode}

\item{limit}{limit of deviation for thresholding associations. Defaults to 0.001}
}
\description{
adduct relationship prediction
}
\examples{
res <- relationshipPredictor(c(65.51148,132.03023,133.03359,168.00691),'n')
}
\author{
Jasen Finch
}

