% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smartsSearch}
\alias{smartsSearch}
\title{SMARTS substructure search}
\usage{
smartsSearch(smile, smart)
}
\arguments{
\item{smile}{a valid SMILE}

\item{smart}{a valid SMARTS symbol}
}
\description{
SMARTS substructure searching for SMILES.
}
\examples{
smartsSearch(amino_acids$SMILES[1],"[OX2H]")
}
