% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{suitableElementRanges}
\alias{suitableElementRanges}
\title{Calculate suitable elemental frequency ranges}
\usage{
suitableElementRanges(mass)
}
\arguments{
\item{mass}{accurate mass}
}
\value{
named numeric vector of element frequencies
}
\description{
Calculate elemental frequency ranges for a given mass which are suitable for molecular formula generation.
}
\examples{
suitableElementRanges(342.11621)
}
