% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formulas.R
\name{transformMF}
\alias{transformMF}
\title{Transform a molecular formula}
\usage{
transformMF(
  MF,
  transformation = "M - [O] + [NH2]",
  transformation_rules_table = transformation_rules()
)
}
\arguments{
\item{MF}{molecular formula to transform}

\item{transformation}{transformation to apply}

\item{transformation_rules_table}{transformations table containing available transformations rules. Defaults to \code{transformation_rules()}.}
}
\value{
A transformed molecular formula.
}
\description{
Transform a molecular formula
}
\details{
\code{NA} will be returned if the molecular formula cannot be transformed.
}
\examples{
transformMF('C4H5O5')
transformMF('C4H5N',transformation = 'M - [OH] + [NH2]')
}
