% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\name{nEntries}
\alias{nEntries}
\alias{nEntries,MetaboliteDatabase-method}
\alias{filterMR}
\alias{filterMR,MetaboliteDatabase-method}
\alias{filterER}
\alias{filterER,MetaboliteDatabase-method}
\alias{filterIP}
\alias{filterIP,MetaboliteDatabase-method}
\alias{filterEntries}
\alias{filterEntries,MetaboliteDatabase-method}
\alias{filterMF}
\alias{filterMF,MetaboliteDatabase-method}
\title{Metabolite database utilities}
\usage{
nEntries(db)

\S4method{nEntries}{MetaboliteDatabase}(db)

filterMR(db, lower, upper)

\S4method{filterMR}{MetaboliteDatabase}(db, lower, upper)

filterER(db, rule)

\S4method{filterER}{MetaboliteDatabase}(db, rule)

filterIP(db, rule)

\S4method{filterIP}{MetaboliteDatabase}(db, rule)

filterEntries(db, IDs)

\S4method{filterEntries}{MetaboliteDatabase}(db, IDs)

filterMF(db, mf)

\S4method{filterMF}{MetaboliteDatabase}(db, mf)
}
\arguments{
\item{db}{S4 object of class MetaboliteDatabase}

\item{lower}{lower mass boundary}

\item{upper}{upper mass boundary}
}
\description{
Filter a MetaboliteDatabase for a given mass range.
}
\examples{
db <- metaboliteDB(amino_acids,descriptors(amino_acids$SMILES))
db <- filterMR(db,100,120)
}
