% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viewAnnotation.R
\name{viewAnnotation}
\alias{viewAnnotation}
\title{View annotation output}
\usage{
viewAnnotation()
}
\description{
A shiny app to view the RData output of the annotateFIE function
}
\author{
Jasen Finch
}

